requirements.txt
Django==2.1.1
PyMySQL==0.9.2
gunicorn==19.9.0
mysqlclient== 1.3.14
djangorestframework== 3.9.0
django-appconf==1.0.2
django-autocomplete-light==3.3.2
django-filter==2.0.0
django-select2==6.3.1
djangorestframework==3.9.0